+++ test/src/SharedMemoryWrite.cpp
@@ -2,11 +2,11 @@
 
 int main()
 {
-    SharedMemory<double> sharedMemory("DOG");
-    std::vector<double> appendData(4 * 4, 44);
+    SharedMemory<long double> sharedMemory("DOG");
+    std::vector<long double> appendData(4 * 4, 44);
     sharedMemory.append(appendData, {1, 4, 4});
 
-    std::vector<double> data(32, 111);
+    std::vector<long double> data(32, 111);
     sharedMemory.append(data, {2, 4, 4});
     
     int temp;
